# Conflicts:
#	src/main/java/application/Application.java
#	src/main/java/service/ClientService.java

+++ src/main/java/service/PlanService.java
@@ -5,5 +5,5 @@ import entity.Plan;
 import java.sql.Date;
 
 public interface PlanService {
-    void createPlan(Integer client_id, Date plan_date_start, Date plan_date_end);
+    void createPlan(Long client_id, Date plan_date_start, Date plan_date_end);
 }
