+++ src/main/java/service/impl/PlanServiceImpl.java
@@ -11,7 +11,7 @@ public class PlanServiceImpl implements PlanService {
     private PlanRepository planRepository;
 
     @Override
-    public void createPlan(Integer client_id, Date plan_date_start, Date plan_date_end) {
+    public void createPlan(Long client_id, Date plan_date_start, Date plan_date_end) {
         planRepository.createPlan(client_id, plan_date_start, plan_date_end);
     }
 }
